+++ pkg/buildfab/container/engine.go
@@ -10,6 +10,7 @@ type Engine interface {
 	PullImage(ctx context.Context, image string) error
 	ImageExists(ctx context.Context, image string) (bool, error)
 	BuildImage(ctx context.Context, config ContainerBuild) (string, error)
+	SlimImage(ctx context.Context, config ContainerSlim) (string, error)
 	
 	// Container operations
 	RunContainer(ctx context.Context, config ContainerConfig) (*ContainerResult, error)